+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## Added
+
+- parallel-gpio: Added `Generic16BitBus`
+- parallel-gpio: Added `PGPIO16BitInterface`
+
 ## [v0.4.1] - 2021-05-10
 
 ### Added